refactor: breakdown the project into folder structure, update README, and rename index.js to server.js
