+++ services/redisService.js
@@ -1,0 +1,17 @@
+import { Redis } from "@upstash/redis";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const redis = new Redis({
+  url: process.env.UPSTASH_REDIS_REST_URL,
+  token: process.env.UPSTASH_REDIS_REST_TOKEN,
+});
+
+export const getCachedResponse = async (key) => {
+  return await redis.get(key);
+};
+
+export const setCache = async (key, value) => {
+  await redis.set(key, value);
+};
